BAT: Add main structure
BAT (Basic Audio Tester) is a simple command line utility intended
to automate audio driver and sound server QA testing with no human
interaction.

BAT works by playing an audio stream and capturing the same stream
in either a digital or analog loopback. It then compares the
captured stream to the original to detemine if the testcase passes
or fails.

BAT supports 4 working modes:
  1. single line playback;
  2. single line capture and analyze;
  3. playback and capture in loop, and analyze captured data;
  4. local analyze without actual playback or capture.
BAT will analyse user input to decide which mode to use.
BAT will create a thread for mode 1, 2 or 3, and will do spectrum
analyse for mode 2, 3 or 4.

Signed-off-by: Lu, Han <han.lu@intel.com>
Signed-off-by: Liam Girdwood <liam.r.girdwood@intel.com>
Signed-off-by: Bernard Gautier <bernard.gautier@intel.com>