BAT: Add spectrum analyze functions
Add functions that detecting signal frequency through spectrum analyze.

The analyze function reads data from wav file, run fft against the data
to get magnitude of frequency vectors, and then calculates the average
value and standard deviation of frequency vectors. After that, we define
a threashold:

    threashold = 3 * standard_deviation + average_value

Frequencies with amplitude larger than threshold will be recognized as
a peak, and the frequency with largest peak value will be recognized as
a detected frequency.

BAT then compares the detected frequency to target frequency, to decide
if the detecting passes or failes.

Signed-off-by: Lu, Han <han.lu@intel.com>
Signed-off-by: Liam Girdwood <liam.r.girdwood@intel.com>
Signed-off-by: Bernard Gautier <bernard.gautier@intel.com>